third_party/WebKit/LayoutTests/platform/linux/external/wpt/preload/download-resources-expected.txt

This is a testharness.js-based test.
FAIL Makes sure that preloaded resources are downloaded assert_equals: resources/CanvasTest.ttf expected 1 but got 0
Harness: the test ran to completion.
